chore: refact, reorganise and minor corrections
