+++ application/app/components/fiches/AccordionCard.tsx
