+++ application/app/components/fiches/ficheEtablissement/ficheEtablissement.tsx
@@ -1,11 +1,11 @@
 import { Etablissement } from '@/app/models/etablissements';
 
-import AccordionCard from './AccordionCard';
-import ActionButtons from './ActionButtons';
+import AccordionCard from '../AccordionCard';
+import ActionButtons from '../ActionButtons';
+import PotentielSolaireCard from '../PotentielSolaireCard';
 import EtablissementCard from './EtablissementCard';
 import GraphiqueCard from './GraphiqueCard';
 import InstallationCard from './InstallationCard';
-import PotentielSolaireCard from './PotentielSolaireCard';
 import ProtectionCard from './ProtectionCard';
 
 interface FicheEtablissementProps {
@@ -23,10 +23,14 @@ export default function FicheEtablissement({ feature }: FicheEtablissementProps)
 			<ActionButtons />
 			<hr className='my-4' />
 			<ProtectionCard isProtected={!!feature.protection} />
-			<PotentielSolaireCard potentiel_solaire={feature.potentiel_solaire} />
+			<PotentielSolaireCard
+				potentiel_solaire={feature.potentiel_solaire}
+				showInterpretation
+				showNbEtablissements={false}
+			/>
 			<hr className='my-4' />
 			<div className='ml-2'>
-				<InstallationCard surface_utile={feature.surface_utile} />
+				<InstallationCard surface_exploitable_max={feature.surface_exploitable_max} />
 				<GraphiqueCard />
 			</div>
 			<hr className='my-4' />
